[Java] System test showing flow control configuration parity between C and Java driver.
